pyproject.toml
[project]
name = "audio-transcription-agent"
version = "1.0.0"
description = "Audio Transcription and Intent Summary Tool with LangChain Agents and MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",

    # Audio processing
    "openai-whisper>=20231117",
    "pydub>=0.25.1",
    "soundfile>=0.12.1",

    # MCP (Model Context Protocol)
    "mcp>=1.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",

    # NLP utilities
    "nltk>=3.8.0",
    "spacy>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

ui = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "websockets>=13.0",
    "jinja2>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
